+++ features/borrow/manage/pipes/manageVaultTransitions.ts
@@ -3,7 +3,7 @@ import { TxHelpers } from 'components/AppContext'
 import { zero } from 'helpers/zero'
 import { Observable } from 'rxjs'
 
-import { WithdrawPaybackDepositGenerateSmartContractLogic } from '../../../../blockchain/calls/proxyActions'
+import { WithdrawPaybackDepositGenerateSmartContractLogicInterface } from '../../../../blockchain/calls/proxyActions/proxyActions'
 import {
   defaultMutableManageVaultState,
   ManageVaultChange,
@@ -263,7 +263,7 @@ export function progressManage(
   txHelpers$: Observable<TxHelpers>,
   state: ManageVaultState,
   change: (ch: ManageVaultChange) => void,
-  proxyActions: WithdrawPaybackDepositGenerateSmartContractLogic,
+  proxyActions: WithdrawPaybackDepositGenerateSmartContractLogicInterface,
 ) {
   const { depositAmount, generateAmount } = state
   const isDepositAndGenerate = depositAmount || generateAmount
